requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
mlflow==2.7.1
minio==7.1.17
pandas
numpy
scikit-learn
boto3
apache-airflow-providers-apache-spark
streamlit
watchdog
plotly
